2. 反思设计模式[Reflection Design Pattern]/2.4 无评分实验-图表生成[Ungraded Lab- Chart Generation]/requirements.txt
# 核心依赖
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=11.3.0
python-dotenv>=1.0.0

# 可选：Jupyter支持
jupyter>=1.0.0
ipython>=8.12.0
